To be able to compile in single precision again
I hope this won't mess up the time printed in the log file.

On branch boil_bojan

modified:   Process/Info_Mod.f90
modified:   Shared/Profiler_Mod.f90
modified:   Shared/Profiler_Mod/Statistics.f90

+++ Sources/Process/Info_Mod.f90
@@ -19,7 +19,7 @@
   type System_Clock_Type
     integer(DP) :: cnt            ! system clock count rate
     integer(DP) :: ini, cur       ! system clock initial and current rate
-    real(DP)    :: wall_time_max  ! maximum wall time
+    real        :: wall_time_max  ! maximum wall time
   end type
 
   !--------------------!
