+++ Sources/Shared/Profiler_Mod/Statistics.f90
@@ -7,7 +7,7 @@
   integer, intent(in)          :: indent     ! 34 for Main_Pro, 1 for Main_con
 !-----------------------------------[Locals]-----------------------------------!
   integer       :: i_fun
-  real(DP)      :: total_time, t_temp, percent_time
+  real          :: total_time, t_temp, percent_time
   integer       :: hours, minutes, seconds
   character(DL) :: line, n_temp
   logical       :: swap
